Move some code into `cli/`
